requirements.txt
anthropic==0.54.0
arrow==1.3.0
backoff==2.2.1
bayesian-optimization==1.5.1
bayespy==0.5.1
biopython==1.84
biopython==1.84
black==24.3.0
bson==0.5.10
coolname==2.2.0
dataclasses-json==0.6.7
datasets==2.1.0
fastai==2.7.17
fastdtw==0.3.4
funcy==2.0
gdcm==1.1
gensim==4.3.3
genson==1.3.0
geopy==2.4.1
google-api-core==2.25.0
google-genai==1.20.0
gym==0.26.2
h5py==3.11.0
hmmlearn==0.3.2
humanize==4.8.0
hyperopt==0.2.7
igraph==0.11.6
imagecodecs==2024.6.1
imbalanced-learn==0.12.3
imgaug==0.4.0
ipykernel==6.29.5
ipython==8.27.0
jsonschema==4.19.2
kornia==0.6.10
librosa==0.10.2.post1
lightgbm==4.5.0
loguru==0.7.2
markovify==0.9.4
matplotlib==3.9.2
nbformat==5.10.4
numba==0.60.0
numpy==1.26.2
omegaconf==2.3.0
openai==1.86.0
opencv-python==4.10.0.84
optuna==4.0.0
pandas==2.1.4
pdf2image==1.17.0
pillow==10.4.0
plotly==5.24.0
pyarrow==17.0.0
pylibjpeg==2.0.1
pyocr==0.8.5
pyparsing==3.1.4
pypdf==4.3.1
pytest==7.4.3
pytorch-lightning==2.4.0
pytz==2024.1
pyyaml==6.0.2
requests==2.31.0
rich==13.7.0
scikit-image==0.24.0
scikit-learn==1.2.2
scikit-learn==1.2.2
scikit-optimize==0.10.2
scipy==1.11.4
scipy==1.11.4
seaborn==0.13.2
sentencepiece==0.2.0
shutup==0.2.0
sklearn-pandas==2.2.0
sympy==1.13.2
tensorpack==0.11
textblob==0.18.0.post0
timm==0.9.7
torch==2.2.0
torchaudio==2.2.0
torchinfo==1.8.0
torchmetrics==1.3.1
torchtext==0.17.0
torchvision==0.17.0
tqdm==4.66.2
tqdm==4.66.2
transformers==4.44.2
xlrd==2.0.1
llama-api-client==0.1.1 #!